ignition.conf in appdata made read/write configurable

+++ include/ignition-entrypoint.sh
@@ -7,12 +7,16 @@ do
 done
 
 #check for the configuration file in the active solution directory
-#set appdata directory
+#set appdata directory if doesn't exist
 if [ ! -d $SNAP_COMMON/solutions/activeConfiguration/ignition ]; then
     mkdir $SNAP_COMMON/solutions/activeConfiguration/ignition
 fi
-cp $SNAP_DATA/data/ignition.conf $SNAP_COMMON/solutions/activeConfiguration/ignition
+if [ ! -f $SNAP_COMMON/solutions/activeConfiguration/ignition/ignition.conf ]; then
+    cp $SNAP_DATA/data/ignition.conf $SNAP_COMMON/solutions/activeConfiguration/ignition
+fi
+#on startup, take configuration from app data area
+cp $SNAP_COMMON/solutions/activeConfiguration/ignition/ignition.conf $SNAP_DATA/data
 
 #execute the ignition script
 echo "Ignition Entrypoint complete, now running main script"
-exec $SNAP/ignition.sh
+exec $SNAP/ignition.sh console
